refactored state and sockets into plugins, attached state to `app` that gets passed around
