--- client/src/includes/state.ts
@@ -1,7 +1,0 @@
-import { reactive } from 'vue'
-import { ClientState } from '../types'
-
-export const state = reactive({
-	static: {},
-	dynamic: {},
-}) as ClientState
