+++ client/src/middleware/authed.ts
@@ -1,6 +1,5 @@
 import { defineMiddleware } from '../includes/functions'
-import { state } from '../includes/state'
 
-export const authed = defineMiddleware((to, from) => {
-	console.log('authed', state)
+export const authed = defineMiddleware((to, from, app) => {
+	console.log('authed', app.state)
 })
