+++ client/src/types/index.ts
@@ -1,4 +1,5 @@
 import type { Systeminformation } from 'systeminformation'
+import { App } from 'vue'
 import {
 	RouteLocationNormalized,
 	RouteLocationRaw,
@@ -19,6 +20,7 @@ type NavigationGuardReturn = void | Error | RouteLocationRaw | boolean
 export type MiddlewareHandler = (
 	to: RouteLocationNormalized,
 	from: RouteLocationNormalized,
+	app: App,
 ) => NavigationGuardReturn | Promise<NavigationGuardReturn>
 
 export interface ComponentMeta extends RouteMeta {
